genes/human/VCP/VCP-bioinformatics/pyproject.toml
[project]
name = "vcp-bioinformatics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "biopython>=1.85",
    "requests>=2.32.5",
]
